refactor: feature/RL-4-app-design move custom hook useWindowWidth to hook folder
