+++ src/components/e-book/ui/pagination/PaginationRanges.tsx
@@ -1,24 +1,14 @@
-import React, { useLayoutEffect, useState } from 'react';
+import React from 'react';
 import Pagination from '@mui/material/Pagination';
 import Stack from '@mui/material/Stack';
+import useWindowWidth from '../../../../hooks/useWindowWidth';
 
 export interface PaginationRangesProps {
   setPage: (val: number) => void;
 }
 
 const PaginationRanges: React.FC<PaginationRangesProps> = ({ setPage }) => {
-  function useWindowWidth() {
-    const [width, setWidth] = useState(0);
-    useLayoutEffect(() => {
-      function updateSize() {
-        setWidth(window.innerWidth);
-      }
-      window.addEventListener('resize', updateSize);
-      updateSize();
-      return () => window.removeEventListener('resize', updateSize);
-    }, []);
-    return width;
-  }
+  const winWith = useWindowWidth();
 
   // **** TO DO ******
   // join  page numbers to original source
@@ -37,9 +27,7 @@ const PaginationRanges: React.FC<PaginationRangesProps> = ({ setPage }) => {
         count={TOTAL_PAGES}
         defaultPage={START_PAGE}
         siblingCount={0}
-        size={
-          useWindowWidth() < 780 ? PageNumberSize.small : PageNumberSize.large
-        }
+        size={winWith < 780 ? PageNumberSize.small : PageNumberSize.large}
         color="primary"
         onChange={(e, page) => setPage(page - 1)}
       />
