+++ src/hooks/useWindowWidth.ts
@@ -1,0 +1,14 @@
+import { useLayoutEffect, useState } from 'react';
+
+export default function useWindowWidth() {
+  const [width, setWidth] = useState(0);
+  useLayoutEffect(() => {
+    function updateSize() {
+      setWidth(window.innerWidth);
+    }
+    window.addEventListener('resize', updateSize);
+    updateSize();
+    return () => window.removeEventListener('resize', updateSize);
+  }, []);
+  return width;
+}
